requirements.txt
# Core Django framework
Django==5.2.3

# JWT authentication
PyJWT==2.8.0

# ASGI server for production (optional)
# uvicorn==0.24.0

# Additional useful packages for Django development
# python-decouple==3.8  # For environment variables management
# Pillow==10.1.0  # For image handling if needed in future
# django-cors-headers==4.3.1  # For CORS if building an API
# requests==2.31.0  # For making HTTP requests if needed

# Development dependencies (uncomment if needed)
# django-debug-toolbar==4.2.0  # For debugging in development
# pytest-django==4.7.0  # For testing
# coverage==7.3.2  # For test coverage
